Move src/draft/ to src/drafting/
